Switch to GNU Make to simplify things
Tup doesn't allow that a command reads from a file it writes (e.g. the
aux file or our Lua module cache), which is otherwise logical, but bad
for us. This is why we had to have a helper runner script to execute
outside of Tup's sight. By ditching Tup, we greatly simplify things.
Moreover, GNU Make is more widely known.