Cache some variables in the use site
Some variables (read: functions) are used only when saving the cache and
only in a single function that is just run once. They thus shouldn't be
cached globally, but inside the function itself. Plus, this way there
are not so many declarations at the start that one has to keep an eye
on.